fixed initializing functions and nullptr on the arduino function call

+++ lib/led/led.h
@@ -3,6 +3,6 @@
 #include <prot.h>
 
 
-const long COLOR = 0xFFFF00;
+const long long COLOR = 0xFFFF00;
 
 void show_unique_led(int led, led_strip& strip);
